Move mangler into the toolchain
